compute: columnated MV sink correction buffer
This commit introduces a "correction v2" buffer that differs from the
existing one in that it stores data in columnated regions that can be
spilled to disk. It follows the general design of the arrangement merge
batcher, with the major difference that it sorts updates by time first,
in an attempt to more efficiently deal with the presence of updates in
the future (commonly introduced by temporal filters).