controller/compute: defer application of early read hold changes
The `ComputeController` provides a sync API for creating collections and
taking out read holds on them, without waiting for the collection to be
created in the `Instance`. Thus the `Instance` might see read hold
changes for collections that haven't been created yet. Stash them for
later application, instead of incorrectly panicking or ignoring the
changes.