controller/compute: sequence external read hold changes
Changes to external `ReadHold`s on compute collections are now sequenced
through the `command_{tx,rx}` channels of the respective `Instance`
controllers. This ensures that `Instance`s observe read hold changes
always after they have observed the creation of the affected collection,
which wasn't the case previously.

As a side effect of this change, each external compute `ReadHold` now
contains a copy of an instance `Client`. These copies keep the connected
`Instance`s alive, even when the `ComputeController` has already dropped
them. This seems fine in that arguably an instance shouldn't shut down
if somebody still has read holds on its collections. But it is a change
to the previous behavior.