ct: don't clone every row when producing `ct_times` (#30123)
In the previous impl, the split in the dataflow graph resulted in every
row being cloned, just to be immediately thrown away. This is wasteful
and potentially doubles the memory requirements. Instead use a custom
1-input 2-output operator (passing through the input).

While I'm in here, also resolve a TODO to switch the operators to
extension traits on `Collection`.

Touches https://github.com/MaterializeInc/database-issues/issues/8427