compute: correctly propagate read holds in shutdown check
Now that some of the read holds held by `Instance` are sequencing their
changes through the command channel, it is no longer guaranteed that an
`apply_read_hold_changes` call also applies cascading changes on
dependencies. The previous `check_empty` check expected this to be the
case, so it would incorrectly fail its asserts.

This is fixed by having the check also drain and apply the contents of
the command channel, until no new read capability changes are produced.
To clarify the semantics, `check_empty` is also renamed to `shutdown`
and explicitly terminates the execution of the `Instance` task.