catalog: Hook up expression cache into startup
This commit hooks up the expression cache into the startup process.
During startup if an expression is found in the cache, then we use that
expression instead of re-optimizing an item. If an expression is not
found in the cache, then we optimize the item and insert the
optimized expression into the cache.

The insertion into the cache happens in bulk as a single operation at
the end of startup instead of once for each object. The reason is that
inserting new expressions may need to invalidate old expressions.
Inserting all expressions in bulk allows us to easily figure out what
should be invalidated vs what shouldn't be invalidated. If we update
the expressions one at a time, then we may accidentally invalidate new
expressions.

Much of the code in this commit involves plumbing things around and
keeping track of what is and isn't cached.

After startup we do not cache newly created items, this is left for
future work. As a result, if environmentd crashes in-between releases
then recovery may be slow if there are many new objects. However,
read-only instances should fully populate the cache during a release.

Works towards resolving #MaterializeInc/database-issues/8384