ct: add an option to start with an initial snapshot of the input (#30092)
This adds a `WITH` option to the `CONTINUAL TASK` syntax (including the
`TRANSFORM USING` sugar) to run the CT statements on an initial snapshot
of the input contents. It is meant to be the default, but there are some
issues to shake out with how it interacts with self-referential CTs
(filed as
https://github.com/MaterializeInc/database-issues/issues/8699).

The motivation is twofold:
- It will probably be common that the user wants this initial backfill.
For example, our system builtin CTs do.
- The initial as_of of a CT is not guaranteed to be strict serializable
(same for MVs). With the initial snapshot, this is not easily
observable, but without it, it's easy to end up in counter-intuitive
situations like: insert, create ct, and then see the insert in the ct
even though it happened "before". It's possible that some users will
want the no snapshot behavior, so leave it in behind an option.

Touches https://github.com/MaterializeInc/database-issues/issues/8427