compute: step back as-ofs of write-only collections
When rehydrating, CT dataflows need to be able to observe changes at
there previous upper (i.e. the current write frontier of the output
shard), to produce the correct output for this time. That means they
need to start reading from their sources (with snapshot excluded) as of
`upper - 1`, and the controller needs to set the dataflow `as_of`
accordingly.

This commit ensures this by (a) adjusting how the compute controller
downgrades read frontiers for write-only collections and (b) the
bootstrap as-of selection process.