compute: fix CT introspection
This commit fixes compute introspection for CT dataflows. Specifically,
before this change the compute controller would report the frontiers of
CT dataflows to introspection. Because CTs also have a storage
collection, they would end up with two entries in `mz_frontiers`. As for
MVs, we need to make the compute controller suppress frontier
introspection for CTs.