ct: fix panic when a CT has a monotonic input
Continual Tasks turn an "input" into diffs by inserting retractions,
which removes any monotonicity properties the collection otherwise would
have had.

The monotonicity information is used in the optimizer, which CTs
currently still share with MVs. This might be a moment for us to fork
out a CT-specific optimizer, but it still doesn't quite seem worth it.

An alternative fix would involve making the optimizer aware of the
difference between CT inputs and normal collections (e.g. we're still
using incorrect stats for CT inputs), but that's a huge change and also
I have no idea how to type it. I consider this already tracked by a
couple of the ct3 TODOs.