Exception for empty batches in the debug validator
As of recently we've started writing down batches that are
"instantaneous", which means that it's possible for two batches to be
totally identical. This has no bad consequences except for triggering
this verification in some tests.

Ideally we could thread through the spine id and preserve uniqueness
here, but since this isn't the essential part of the validation, it's
much easier to just check that the duplicate batch is of the expected
sort.