repr: Remove outdated jsonb packer comments
When the `JsonbPacker` was first implemented in
c8fa525c6df8bad86a46353208743d2c06466d92, the `JsonbPacker` would
serialize integers by first converting them to an `f64`. If the integer
didn't fit into an `f64`, then an error was returned. Many of the Jsonb
serialize and deserialize methods include a comment of the form:

  /// Errors if any of the contained integers cannot be represented exactly as
  /// an [`f64`].

Later b7b1816eb9eeceafbf79603be4db1de5a3d35ba3, updated the
serialization format to store all numbers as numerics, which have
enough precision for all accepted integer types. As a result, the
comment about integers and `f64` was no longer accurate, however, the
comment was not removed.

This commit removes the outdated comment about integers and `f64`. It's
very possible that some of the `JsonbPacker` methods can never return
an `Err`. However, verifying this and updating the APIs is left as
future work.