storage: in feedback upsert, allow emitting partial updates
This allows ingesting updates at a given timestamp while the timestamp
is not yet complete.

For this to work we need to modify both ingestions sides:
- ingesting updates from the source input
- ingesting finalized state updates from the persist input

For the source input, we introduce _provisional values_ in upsert state.
These are valid only for the timestamp at which they are written. For
other timestamps we use the _finalized value_ in state. Keeping the
provisional value allows emitting updates for a timestamp that we might
retract again, within that timestamp. Provisional updates are
overwritten when ingesting updates from the persist side.

For the persist input, we need to be able to ingest partial updates, in
arbitrary order. For this, we extend the "snapshot chunk consolidation"
to work for arbitrary inputs, not just the first snapshot from the
persist input. This has the nice side effect of simplifying the persist
ingestion side because we now have only one method of ingesting updates.