ct: add CREATE CONTINUAL TASK .. RETAIN .. WHILE (<expr>)
Like `TRANSFORM .. USING`, this desugars into a normal CT, but with a
restricted "shape". In this case, to:

- `INSERT INTO name SELECT * FROM input WHERE <expr>`
- `DELETE FROM name WHERE NOT <expr>`

This is primarily targeted at temporal filters (otherwise it's
sufficient to filter only on the insert and `TRANSFORM .. USING` can be
used). However, a temporal filter on append-only data is one of the most
existing initial CT use cases, so feels worth it.

It feels a bit bad that we can't combine these into a generalized
transform using, but:
- We're not committed to anything here yet. At the very least, we'll go
  through full sql-council approval before moving any of this to prpr.
  At the moment, we're just looking to get something demo-able so Field
  Eng can shake us up some design partners.
- The reason to have the general syntax is because it's composable and
  regular, and the reason to have sugar is because the fully general
  syntax is verbose for simple use cases, but potentially no need for
  the sugar to be composable.
- It's hard to answer what a CT even is. It might be possible that it's
  easier to "teach" them by starting with a set of targeted sugars.