ct: fix self-reference + with snapshot
Previous to this, things were in a weird state where the CT since would
do an initial CaA of `[0, A+1)` on the output. `[0, A)` would be correct
for "with snapshot" but changing it to this broke self-references.

I left a big explanation in the continual_task module rustdoc, but the
fix comes down to:
- Ignore self-references when deciding if a dataflow's input are
  available for scheduling.
- For self-references only, use `A-1` as the as_of for the
  `suppress_early_progress` operator.
- Change the initial CaA to `[0, A)`, which is now correct for all
  cases.