Improve `Equivalences::union` implementation performance (#30134)
This PR improves the performance of `Equivalences::union`, and
potentially alters its outputs (on non-canonical inputs).

Superficially, there is now a method `union_many` that accepts multiple
equivalences to merge, and which `union` just calls. This allows our
variadic `Union` stages to call this once, rather than a sequence of
rebuilds. This also reduces a fair bit of cloning. I did not measure how
impactful this step is in isolation.

More importantly, the implementation is changed from quadratic all-pairs
set intersection to a map based approach that intersects the keys of two
maps, and builds equivalence classes from the results. If an expression
maps to `foo` and `bar` under the two maps, then a class `(foo, bar)` is
created, and assigned a new class identifier. This new identifier is
used as `foo` in the next iteration, with the next input equivalence
class, repeating through each of the classes.

This ends up being linear time, which is great.

This does not improve the inherent limitation that we are looking for
syntactic equality, so `f(#0)` and `f(#1)` will not match, even if in
the first input `#0 == #1`. We might need something more like e-graphs
(or our `project` implementation, which is a simplified form of
intersection) to get this.

This PR has the potential to *change* the output when the input is not
an equivalence relation. It would previous search through
non-equivalence class `self.classes`, firing on hits in those lists.
When `classes` describes an equivalence relation we'll get the same
output, but when they are e.g. not closed under transitivity, .. I don't
know what output we are producing.

### Motivation

<!--
Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug.

    [Ensure issue is linked somewhere.]

  * This PR adds a known-desirable feature.

    [Ensure issue is linked somewhere.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

[Write a brief specification for the feature, including justification
for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

[Describe what was wrong with the existing code, if it is not obvious.]
-->

### Tips for reviewer

<!--
Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
* [Some function/module/file] is pure code movement and only needs a
skim.

Delete this section if no tips.
-->

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.