compute: make as-of back-stepping a soft constraint
Stepping back the as-of for dataflows with persisted outputs relies on
the compute controller holding back input collections for these
dataflows accordingly. This is currently not guaranteed during a release
upgrade, as the holding-back was introduced at the same time as the
stepping-back in as-of selection.

To avoid panicking in upgrade tests, this commit makes the stepping-back
a soft as-of constraint. It is still required for the correctness of CTs
so we need to revert this commit in a follow-up. This should be safe to
do after at least one Mz version was released with the new holding-back
behavior, and will definitely be safe to do once the old holding-back
behavior has rotated out of our upgrade tests.