ct: fix bug when creating TRANSFORM USING in cluster other than u1
We weren't correctly storing the `IN CLUSTER` clause in the catalog
`create_sql`, which resulted in the catalog thinking the CT belonged to
whatever cluster was default in `parse_item`. This only happened for the
`TRANSFORM USING` sugar.

Fix by adding parser support for `IN CLUSTER` to TRANSFORM USING (seems
good, anyway!).