ct: fix bug where CT name was not escaped in `create_sql`
This happens because, unlike other things, the CT name is represented on
the statement as a `T::ItemName` instead of `UnresolvedItemName` (so
that we can sniff out the assigned LocalId). There are probably other
ways to fix this, but I don't have enough context to have an opinion on
which is the best.