catalog: Split local and global expression cache
This commit separates the `Expression` struct, which is cached in the
expression cache, into two separate structs. The `LocalExpressions`
struct stores a local optimization of an expression and the
`GlobalExpressions` struct stores a global optimization of an
expression. Each expression type is stored as a separate entry in the
cache.

Certain object types only have a local optimization in the catalog,
others only have a global optimization, and others have both
representations. Storing the different expression types separately
makes working with the different object types easier.

While we're changing the cache structure, we also remove the `notices`
field, which is actually just a denormalized view of the
`dataflow_metainfos` field.

Works towards resolving #MaterializeInc/database-issues/8384