catalog: Fix builtin view migrations
When a builtin view is migrated and 0dt is enabled, we update the
view's fingerprint stored in the durable catalog. This update is
materialized in the catalog as a retraction of the old fingerprint
followed by an addition of the new fingerprint. For all builtin item
types except for views, we have special logic to identify this update
and treat it as a no-op. This is important because in memory the items
maintain denormalized state, such as dependents, which would be
expensive to rebuild. Builtin views have a special code path for
handling retractions and additions that is missing this no-op logic.
Furthermore, that code path doesn't properly rebuild the denormalized
state which can lead to incorrect metadata stored in memory.

This commit fixes the issue described above by adding the no-op logic
to the special code path for builtin views. The new logic is model
after the existing logic, which can be found in
`CatalogState::apply_system_object_mapping_update`.