catalog: Optimize catalog opens (#30335)
An unopened catalog caches the fence token, configs collection, and
settings collection. Previously, in order to do this it would clone all
catalog updates, attempt to deserialize them, and then check if it was
one of those three cached update kinds. This was hugely wasteful
because the cached updates makes up a tiny portion of all updates.

This commit fixes the issue, by first peeking into the JSON of the
serialized updates, and only deserializing the updates that are one of
the cached updates. The process of peeking into the JSONB is a bit
scary and may break some abstraction boundaries, but is hopefully worth
it for the performance improvements.

Works towards resolving #MaterializeInc/database-issues/8384