storage-client: fix frontier watching
The previous implementation of the shard frontier watcher task in
`StorageCollections` could skip over frontier advancements. This is
particularly problematic if the skipped advancement was to the empty
frontier because the storage controller would never learn about the fact
that the collection had reached the empty frontier in this case.