catalog: Improve serializability of expressions
This commit adds the necessary logic so that all cached expressions can
be serialized by serde. That includes:

- Serializing/deserializing map keys through a String.
- Adding `Serialize` and `DeserializeOwned` trait bounds to various
  structs.
- Adding the ability for regexes to deserialize owned strings.

Many of the serailization errors that existed before this commit would
only manifest as runtime errors, not compile time errors. To help catch
more errors, this commit also implements `Arbitrary` on most of the
cached expressions. The one exception is
`DataflowDescription<mz_compute_types::plan::Plan>`, which is saved for
future work. The `Arbitrary` implementations are fairly slow, which
causes all expression cache tests to be slow.

Additionally, if serialization/deserialization fails, then we log an
error and move on instead of panic. The cache is purely an
optimization, so we don't want some runtime error to bring down all of
environmentd.

Works towards resolving #MaterializeInc/database-issues/8384