persist_source: initialize read handle before waiting
Currently it can happen that the controller fails to hold back the
sinces of dataflow inputs to their as-of times (database-issues#8729).
If that happens, waiting for the start signal before obtaining a read
handle in `persist_source`s makes panics signficantly more likely, as it
increases the opportunity for persist to compact the shard the
`persist_source` is planning to read from.