adapter: Don't create new builtins in read-only
This commit skips creating new builtin objects in read-only mode. While
in read-only mode, we cannot allocate new `GlobalId`s durably. So
there's no way to create a new builtin object with a `GlobalId` and
guarantee that we'll use the same `GlobalId` when rebooting in
read-write mode.

Fixes #MaterializeInc/database-issues/8731