catalog: Implement Arbitrary on expression types (#30369)
This commit adds an Arbitrary implementation on most of the structs
used in the expression cache. The one exception is
`DataflowDescription<mz_compute_types::plan::Plan>`, which is saved for
future work.

This has helped us identify existing issues with the serialization
requiring
us to ignore some of the existing tests until it is fixed.

Additionally, if serialization/deserialization fails in production,
then we log an error and move on instead of panic. The cache is purely
an optimization, so we don't want some runtime error to bring down all

Works towards resolving #MaterializeInc/database-issues/8384