catalog: Fix expression cache serialization
This commit fixes the serialization of optimized expressions in the
expression cache by switching from JSON to `bincode`. JSON
serialization had the following issues:

- All maps needed to serialize their keys through strings.
- Serialization issues tended to manifest as runtime errors instead
  of compile time errors.
- `bincode` and JSON don't get along.

Furthermore, compute is already serializing many of these expressions
using `bincode`, so we know it mostly works properly.

Unfortunately, this makes it more difficult to read and debug the
expression cache. We don't anticipate this being a huge issue, because
we can always disable or clear the expression cache if an issue arises.

Works towards resolving #MaterializeInc/database-issues/8384