storage: optional probe streams from source operators
This commit introduces the ability for source operators to omit
returning a probe stream. This is appropriate for sources that don't
support probing the upstream system, such as load generators (which
don't have an upstream system).

The "latest" reclocking mode still expects to receive probes, so the
reader pipeline synthesizes a probe stream from source progress events
and uses that in cases where the source doesn't support probing. This is
the same approach previously used by the load generators, just
centralized in a single place.