Disable the new instrospection sources
This commit disables the newly added introspection sources
`mz_compute_dataflow_global_ids_per_worker` and
`mz_compute_lir_mapping_per_worker`, because Mz's 0dt mechanism is
currently not prepared to handle new introspection sources.

"Disabling" here means removing the registration of the new `Log`
builtins and the `View`s depending on them. The intention is to keep the
diff as small as possible, to make this change less risky.