storage: enable changing reclock mode without restart
This commit enables changing the source reclock strategy, through the
`storage_reclock_to_latest` flag, of a source without having to restart
the source dataflow. The intention is to derisk the production rollout
by having a way to quickly revert to the old behavior without causing
downtime.

This change mostly consists of moving the check of the dyncfg into the
remap loop. However, changing the reclock mode at runtime can result in
a temporary regression of the binding ts, so the remap operator now also
has to deal with that possibility and ensure that no regressing bindings
are minted.