compute: properly initialize uppers of storage sinks
This commit makes the compute controller advance the write frontier of
new storage sink collections (MVs, CTs) to the write frontier of the
target storage collection.

Previously it didn't do that, which is normally fine because once a
replica has installed the dataflow it will send a `Frontiers` response
to bump the write frontier in the controller. However, if the cluster
has no replicas (e.g. because it is paused between MV refreshes) this
doesn't happen. In this case, the write frontier would get stuck at
dataflow creation time, which made the implied capability get stuck at
the same time too, which in turn held back compaction of the dataflow
inputs to the same time. If the write frontier of the target storage
collection is significantly in the future, this holding back is
undesired, as it prevents compaction of the inputs more than necessary.